+++ src/main/java/com/gistpetition/api/verification/application/password/FindPasswordValidator.java
@@ -1,4 +1,4 @@
-package com.gistpetition.api.verification.application;
+package com.gistpetition.api.verification.application.password;
 
 public interface FindPasswordValidator {
     void checkIsVerified(String username, String verificationCode);
