+++ src/main/java/com/gistpetition/api/verification/application/password/FindPasswordValidatorImpl.java
@@ -1,25 +1,24 @@
-package com.gistpetition.api.verification.application;
+package com.gistpetition.api.verification.application.password;
 
 import com.gistpetition.api.exception.verification.InvalidVerificationInfoException;
 import com.gistpetition.api.exception.verification.NoSuchVerificationCodeException;
 import com.gistpetition.api.exception.verification.NotConfirmedVerificationCodeException;
-import com.gistpetition.api.verification.domain.SignUpVerificationInfoRepository;
+import com.gistpetition.api.verification.domain.PasswordVerificationInfoRepository;
 import com.gistpetition.api.verification.domain.VerificationInfo;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
 
+@RequiredArgsConstructor
 @Component
-public class SignUpValidatorImpl implements SignUpValidator {
-    private final SignUpVerificationInfoRepository signUpVerificationInfoRepository;
+public class FindPasswordValidatorImpl implements FindPasswordValidator {
 
-    public SignUpValidatorImpl(SignUpVerificationInfoRepository signUpVerificationInfoRepository) {
-        this.signUpVerificationInfoRepository = signUpVerificationInfoRepository;
-    }
+    private final PasswordVerificationInfoRepository passwordVerificationInfoRepository;
 
     @Override
     public void checkIsVerified(String username, String verificationCode) {
-        VerificationInfo verificationInfo = signUpVerificationInfoRepository.findByUsernameAndVerificationCode(username, verificationCode)
+        VerificationInfo verificationInfo = passwordVerificationInfoRepository.findByUsernameAndVerificationCode(username, verificationCode)
                 .orElseThrow(NoSuchVerificationCodeException::new);
 
         if (!verificationInfo.isConfirmed()) {
