+++ src/main/java/com/gistpetition/api/verification/application/signup/SignUpValidator.java
@@ -1,4 +1,4 @@
-package com.gistpetition.api.verification.application;
+package com.gistpetition.api.verification.application.signup;
 
 public interface SignUpValidator {
     void checkIsVerified(String username, String verificationCode);
