+++ src/test/java/com/gistpetition/api/user/UserServiceTest.java
@@ -13,7 +13,7 @@ import com.gistpetition.api.user.dto.request.SignUpRequest;
 import com.gistpetition.api.user.dto.request.UpdatePasswordRequest;
 import com.gistpetition.api.user.dto.request.UpdateUserRoleRequest;
 import com.gistpetition.api.utils.password.Encoder;
-import com.gistpetition.api.verification.application.SignUpValidator;
+import com.gistpetition.api.verification.application.signup.SignUpValidator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
